+++ hasklepias-core/src/Hasklepias/ReexportsUnsafe.hs
@@ -16,6 +16,7 @@ module Hasklepias.ReexportsUnsafe
     module Data.Aeson
   , module GHC.Exts
   , module GHC.IO
+  , module Lens
   , module Test.Tasty
   , module Test.Tasty.HUnit
   ) where
@@ -25,6 +26,8 @@ import           Data.Aeson                     ( ToJSON(..)
                                                 )
 import           GHC.Exts                       ( IsList(fromList) )
 import           GHC.IO                         ( IO(..) )
-
+import           Data.Generics.Internal.VL.Lens as Lens ( (^.) ) 
+import           Data.Generics.Product as Lens         ( HasField(field) ) 
+import           Data.Generics.Sum   as Lens           ( AsAny(_As) ) 
 import           Test.Tasty              hiding ( after )
 import           Test.Tasty.HUnit