+++ edm/src/EventData/Context/Domain/Enrollment.hs
@@ -13,17 +13,18 @@ module EventData.Context.Domain.Enrollment(
 
 import Data.Aeson               ( FromJSON(..) )
 import Data.Eq                  ( Eq )
+import Data.Maybe               ( Maybe(..) )
 import GHC.Generics             ( Generic )
 import GHC.Show                 ( Show )
 import EventData.Context.Facts.Plan ( Plan, emptyPlan )
 
 -- | Enrollment
 newtype EnrollmentFacts = EnrollmentFacts {
-     plan :: Plan 
+     plan :: Maybe Plan 
   } 
   deriving( Eq, Show, Generic )
 
 emptyEnrollmentFact :: EnrollmentFacts
-emptyEnrollmentFact = EnrollmentFacts emptyPlan
+emptyEnrollmentFact = EnrollmentFacts Nothing
 
 instance FromJSON EnrollmentFacts where