+++ hasklepias-core/ChangeLog.md
@@ -3,6 +3,7 @@
 ## 0.22.5
 
 * Adds `asum` to reexports module.
+* Reexports a few lens-related functions.
 
 ## 0.22.4