+++ hasklepias-core/ChangeLog.md
@@ -1,5 +1,10 @@
 # Changelog for hasklepias-core
 
+## 0.22.5
+
+* Adds `asum` to reexports module.
+* Reexports a few lens-related functions.
+
 ## 0.22.4
 
 * Updates `interval-algebra` dependency to 1.1.0.